requirements.txt
langchain>=0.2
langgraph>=0.2
chromadb>=0.5
rank-bm25>=0.2
sentence-transformers>=3.0
pypdf>=4.2
python-dotenv>=1.0
pydantic>=2.7
pydantic-ai>=0.0.16
fastapi>=0.111
uvicorn[standard]>=0.30
google-generativeai>=0.7
langchain-google-genai>=2.0
